Refactor form controllers into controllers module
